requirements.txt
-e .
coverage
django-stubs
flake8
mypy
pre-commit
pytest
yapf
